requirements.txt
alabaster==0.7.10
appnope==0.1.0
asn1crypto==0.24.0
astroid==1.6.4
atari-py==0.1.1
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.6.0
backcall==0.1.0
bleach==2.1.3
bokeh==0.12.16
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
colorama==0.3.9
cryptography==2.2.2
cycler==0.10.0
Cython==0.28.3
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
Flask==1.0.2
flatbuffers==2015.12.22.1
funcsigs==1.0.2
future==0.16.0
gym==0.10.5
h5py==2.8.0
html5lib==1.0.1
idna==2.6
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.2.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.3
numpydoc==0.8.0
opencv-python==3.4.1.15
packaging==17.1
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.1
patsy==0.5.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prompt-toolkit==1.0.15
protobuf==3.5.2
psutil==5.4.5
ptyprocess==0.5.2
py==1.5.3
pycodestyle==2.4.0
pycparser==2.18
pyflakes==2.0.0
pyglet==1.3.2
Pygments==2.2.0
pylint==1.9.1
PyOpenGL==3.1.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.6.1
python-dateutil==2.7.3
python-snappy==0.5.2
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.2
ray==0.4.0
redis==2.10.6
requests==2.18.4
rope==0.10.7
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
statsmodels==0.9.0
tensorflow==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
